+++ src/ts/components/volumetogglebutton.ts
@@ -9,10 +9,14 @@ export class VolumeToggleButton extends ToggleButton<ToggleButtonConfig> {
   constructor(config: ToggleButtonConfig = {}) {
     super(config);
 
-    this.config = this.mergeConfig(config, {
+    const defaultConfig: ToggleButtonConfig = {
       cssClass: 'ui-volumetogglebutton',
       text: 'Volume/Mute',
-    }, this.config);
+      onClass: 'muted',
+      offClass: 'unmuted',
+    };
+
+    this.config = this.mergeConfig(config, defaultConfig, this.config);
   }
 
   configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
@@ -21,17 +25,29 @@ export class VolumeToggleButton extends ToggleButton<ToggleButtonConfig> {
     let muteStateHandler = () => {
       if (player.isMuted()) {
         this.on();
+
+        // When the volume is unmuted and the volume level is veeeery low, we increase it to 10%. This especially helps
+        // in the case when the volume is first turned down to 0 and then the player is muted; when the player gets
+        // unmuted it would switch to volume level 0 which would seem like unmuting did not work, and increasing the
+        // level a bit helps to overcome this issue.
+        if (player.getVolume() < 10) {
+          player.setVolume(10);
+        }
       } else {
         this.off();
       }
     };
 
     let volumeLevelHandler = () => {
-      // Toggle low class to display low volume icon below 50% volume
-      if (player.getVolume() < 50) {
-        this.getDomElement().addClass(this.prefixCss('low'));
-      } else {
-        this.getDomElement().removeClass(this.prefixCss('low'));
+      const volumeLevelTens = Math.ceil(player.getVolume() / 10);
+      this.getDomElement().data(this.prefixCss('volume-level-tens'), String(volumeLevelTens));
+
+      // When the volume is turned down to zero, switch into the mute state of the button. This avoids the usability
+      // issue where the volume is turned down to zero, the button shows the muted icon but is not really unmuted, and
+      // the next button press would switch it into the mute state, visually staying the same which would seem like
+      // an expected unmute did not work.
+      if (volumeLevelTens === 0) {
+        this.off();
       }
     };