+++ src/ts/player.d.ts
@@ -54,11 +54,21 @@ declare namespace bitmovin {
   interface PlayerAPI {
     /**
      * Subscribes an event handler to a player event.
+     * This will be replaced by {@link PlayerAPI.on} in version 8
      *
      * @param eventType The type of event to subscribe to.
      * @param callback The event callback handler that will be called when the event fires.
      */
     addEventHandler(eventType: PlayerAPI.EVENT, callback: PlayerAPI.PlayerEventCallback): PlayerAPI;
+    /**
+     * Subscribes an event handler to a player event.
+     *
+     * @param eventType The type of event to subscribe to.
+     * @param callback The event callback handler that will be called when the event fires.
+     *
+     * @since v7.7
+     */
+    on(eventType: PlayerAPI.EVENT, callback: PlayerAPI.PlayerEventCallback): PlayerAPI;
     /**
      * Sends custom metadata to Bitmovin's Cast receiver app.
      *
@@ -323,12 +333,22 @@ declare namespace bitmovin {
     play(issuer?: string): PlayerAPI;
     /**
      * Removes a handler for a player event.
+     * This will be replaced by {@link PlayerAPI.off} in version 8
      *
      * @param eventType The event to remove the handler from
      * @param callback The callback handler to remove
      */
     // TODO remove string type option (this is a temporary hack for PlayerWrapper#clearEventHandlers)
     removeEventHandler(eventType: PlayerAPI.EVENT | string, callback: PlayerAPI.PlayerEventCallback): PlayerAPI;
+    /**
+     * Removes a handler for a player event.
+     *
+     * @param eventType The event to remove the handler from
+     * @param callback The callback handler to remove
+     *
+     * @since v7.7
+     */
+    off(eventType: PlayerAPI.EVENT | string, callback: PlayerAPI.PlayerEventCallback): PlayerAPI;
     /**
      * Removes the existing subtitle/caption track with the track ID specified by trackID. If the track is
      * currently active, it will be deactivated and then removed. If no track with the given ID exists,